# docs/static/openapi.yaml
openapi: 3.0.0
info:
  contact:
    name: BearMount
    url: https://github.com/WhispersOfJ/bearmount/issues
  description: REST API for BearMount — Usenet WebDAV server with NZB queue,
    health monitoring, ARR integration, FUSE/rclone mounting, and Stremio addon.
  license:
    name: MIT
  termsOfService: https://github.com/WhispersOfJ/bearmount
  title: BearMount API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      description: Authenticates with username and password and returns a JWT access
        token. Rate-limited to 10/min per IP.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api.LoginRequest"
        description: Login credentials
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/api.APIResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/api.AuthResponse"
                    type: object
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      summary: Login
      tags:
        - Auth
  /auth/register:
    post:
      description: Creates the first admin user account. Only allowed when no users
        exist yet.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api.RegisterRequest"
        description: Registration details
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/api.APIResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/api.AuthResponse"
                    type: object
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      summary: Register
      tags:
        - Auth
...
  /auth/reset-admin-password:
    post:
      description: Resets the password for a direct-auth user. Only available when
        login is disabled.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api.ResetAdminPasswordRequest"
        description: Reset credentials
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      summary: Reset admin password
      tags:
        - Auth
...
  /user:
    get:
      description: Returns information about the currently authenticated user.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/api.APIResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/api.UserResponse"
                    type: object
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Get current user
      tags:
        - User
  /user/api-key/regenerate:
    post:
      description: Generates a new API key for the authenticated user, invalidating
        the old one.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/api.APIResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/api.UserResponse"
                    type: object
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Regenerate API key
      tags:
        - User
  /user/logout:
    post:
      description: Invalidates the current session and JWT token.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Logout
      tags:
        - User
  /user/password:
    put:
      description: Changes the password for the currently authenticated user.
      requestBody:
        content:
          application/json:
            schema:
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
              type: object
        description: Password change request
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Change password
      tags:
        - User
  /user/refresh:
    post:
      description: Issues a new JWT access token using the current valid token.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/api.APIResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/api.AuthResponse"
                    type: object
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Refresh token
      tags:
        - User
...
servers:
  - url: //localhost:8080/api
components:
...
  securitySchemes:
    ApiKeyAuth:
      in: query
      name: apikey
      type: apiKey
  schemas:
    api.APIError:
      properties:
        code:
          type: string
        details:
          type: string
        message:
          type: string
      type: object
    api.APIMeta:
      properties:
        count:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
      type: object
    api.APIResponse:
      properties:
        data: {}
        error:
          $ref: "#/components/schemas/api.APIError"
        meta:
          $ref: "#/components/schemas/api.APIMeta"
        success:
          type: boolean
      type: object
...
    api.AuthResponse:
      properties:
        message:
          type: string
        redirect_url:
          type: string
        user:
          $ref: "#/components/schemas/api.UserResponse"
      type: object
...
    api.LoginRequest:
      properties:
        password:
          type: string
        username:
          type: string
      type: object
...
    api.RegisterRequest:
      properties:
        email:
          type: string
        password:
          type: string
        username:
          type: string
      type: object
    api.ResetAdminPasswordRequest:
      properties:
        new_password:
          type: string
        username:
          type: string
      type: object
...
    api.UserResponse:
      properties:
        api_key:
          type: string
        avatar_url:
          type: string
        email:
          type: string
        id:
          type: string
        is_admin:
          type: boolean
        last_login:
          type: string
        name:
          type: string
        provider:
          type: string
      type: object
...